% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaChoice.R
\name{lambdaChoice}
\alias{lambdaChoice}
\title{lambdaChoice}
\usage{
lambdaChoice(
  data,
  name.exposure,
  name.outcome,
  name.mediators,
  name.covariables = NULL,
  weights = NULL,
  groups = NULL,
  lambdamax = 50,
  N1 = 10,
  selectedMin = 0.3 * length(outcome),
  selectedMax = 0.6 * length(outcome),
  L0 = 1,
  eta = 2,
  epsilon = 0.001
)
}
\arguments{
\item{data}{a data.frame with the exposure, the outcome and mediators.}

\item{name.exposure}{a character string indicating the name of the exposure(s) in data.}

\item{name.outcome}{a character string indicating the name of the outcome in data.}

\item{name.mediators}{a character string indicating the name of mediators in data.}

\item{name.covariables}{a character string indicating the name of mediators in data.}

\item{weights}{writing in progress}

\item{groups}{writing in progress}

\item{lambdamax}{writing in progress}

\item{N1}{writing in progress}

\item{selectedMin}{writing in progress}

\item{selectedMax}{writing in progress}

\item{L0}{writing in progress}

\item{eta}{writing in progress}

\item{epsilon}{writing in progress}
}
\value{
lambdaChoice returns a list which contains the lambda to use in mahi function.
}
\description{
'lambdaChoice' is used to select mediators trough a high number of mediators and estimate various quantities for causal mediation analysis using "multimediate", with the selected mediators.
}
