% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahi.R
\name{mahi}
\alias{mahi}
\title{mahi}
\usage{
mahi(
  data,
  name.exposure,
  name.outcome,
  name.mediators,
  name.covariables = NULL,
  lambda = NULL,
  Nboot = 100,
  L0 = 0.1,
  eta = 2,
  weights = NULL,
  groups = NULL,
  epsilon = 0.001,
  Kmax = NULL,
  bin = FALSE,
  dostep2 = FALSE,
  p.adjust.method = "bonferroni",
  pvalseuil = 0.05
)
}
\arguments{
\item{data}{a data.frame with the exposure, the outcome and mediators.}

\item{name.exposure}{a character string indicating the name of the exposure in data.}

\item{name.outcome}{a character string indicating the name of the outcome in data.}

\item{name.mediators}{a character string indicating the name of mediators in data.}

\item{name.covariables}{a character string indicating the name of mediators in data.}

\item{lambda}{vector or a single value of lambda use for group-lasso.}

\item{Nboot}{number of bootstrap for stability selection of mediators. Default is 100.}

\item{L0}{writing in progress}

\item{eta}{writing in progress}

\item{weights}{writing in progress}

\item{groups}{writing in progress}

\item{epsilon}{writing in progress}

\item{Kmax}{maximum of mediators keept after the ranking obtained with stability selection.}

\item{bin}{a logical value indicating if the outcome is binary. if 'TRUE' a probit regression will be use is the second step. Default is 'FALSE'.}

\item{dostep2}{a logical value indicating if the a multiple test using the p-value calculated with multimediate (a multiple mediation analysis) have to be made. Default is 'FALSE'.}

\item{p.adjust.method}{correction method to use on the p-value calculated with multimediate among c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY",}

\item{pvalseuil}{p-value threshold for multiple testing. Default is 0.05.}
}
\value{
mahi returns an object of class "mahi", a list that contains the components listed below.
}
\description{
'mahi' is used to select mediators trough a high number of mediators and estimate various quantities for causal mediation analysis using "multimediate", with the selected mediators.
}
\examples{
# example code

#data(data_cont)

#lambda <- lambdaChoice(data_cont,name.exposure="Exposure",name.outcome=
#"Outcome",name.mediators=paste0("M", 1:500),lambdamax=30600,selectedMin=50,
#selectedMax=100)

#mahifit <- mahi(data=data_cont,name.exposure="Exposure",name.outcome=
#"Outcome",name.mediators=paste0("M", 1:500),
#lambda=lambda,Nboot=30,L0=.1,eta=2,epsilon=.001, Kmax=30,bin=FALSE,
#dostep2=TRUE,p.adjust.method="hochberg",pvalseuil=0.05)

##The function mahi returns a list containing several results such as the
## mediators selected in the first step.
#summary(mahifit)

##To see the mediators selected after the step 1, use :
#   mahifit$Kmaxrank

## Remember that the first 30 candidates are the real mediators. The first
## 10 (1 to 10) have a strong mediated effect, the next 10 (11 to 20) have a
## medium mediated effect, and the other 10 (21 to 30) have a very
## weak effect.

##To see the mediators count for each mediator after the stability selection
## (step 1), use :
#  mahifit$bootcount

##To see the ranking after the stability selection (step 1), use :
#mahifit$ranking

##To see the results of the multiple analysis with mediators selected at
## the step 1, use :
#summary(mahifit$multimed)

## You can see the p-value in the previous table or use :
#   mahifit$pvals

## The corrected p-value for the multiple test obtained with method given
## can be disp with :
#mahifit$pvalscorr

##It is then possible to make another correction for a multiple test with the
## mediators selected in step 1.
##To see the mediators selected after step 2 you can use these lines of
# codes.

#mahifit$step2
}
